// src/front/src/api/rcp/models/dayEntryResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Rcp API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export interface DayEntryResponse {
	/** @nullable */
	gid?: string | null;
	/** @nullable */
	workDate?: string | null;
	/** @nullable */
	startTime?: string | null;
	/** @nullable */
	endTime?: string | null;
	workMinutes?: number;
	hours?: number;
	minutes?: number;
	/** @nullable */
	timeFormatted?: string | null;
	decimalHours?: number;
	/** @nullable */
	notes?: string | null;
}